requirements/requirements_local.txt
-r requirements_production.txt

# Development Tools
coverage==7.4.1
black==24.1.1
flake8==7.0.0
pytest==8.0.0
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.5.1
django-debug-toolbar==4.2.0
ipython==8.20.0
mypy==1.8.0
django-stubs==4.2.7
